Can add location data by hand
This commit adds the code and templates needed to input user location
data by hand.  The display of the data needs to be changed to deal with
the time stamps being in UTC and to allow for more than the newest
location (perhaps the last N locations or all locations in a date
range).

Signed-off-by: Eric B Munson <emunson@mgebm.net>