Add http api
In order to make this all useful and to satisfy the whoel reason I got
started, we need to make this data consumable by an Android application.
The easiest way I could think of is expose an API.

With the API as it stands now, you can authenticate, change who can read
your location data, change your email or password, lookup users, and
fetch location data for users.

Signed-off-by: Eric B Munson <emunson@mgebm.net>