Drop the use of timezone
Becuase it doesn't seem to be available on my server, let s drop the
use of timezone aware datetime objects.  Instead, the documentation for
the api will need to be clear that it is up to each client to localize
the time before displaying it and convert any time to UTC before sending
it.

A future commit will add a Timezone parameter to each user which will
be used to convert times before they are displayed online (times sent
out via the API will never be converted).

Signed-off-by: Eric B Munson <emunson@mgebm.net>